vge/materials/phong: fix build on Linux
This commit fixes the build on Linux.

Firstly, it removes the `.exe` extension from the glslangValidator
tool. Hopefully, this should still work on Windows. @lakal3 would
you be able to verify? I don't have access to a Windows machine.

It also updates the name of the SPIR-V files used by Go embed.
I reckon the `.gls` extension was an artifact from an earlier
version of VGE?