cpp: make Vulkan SDK path optional (default set to empty string)
On Arch Linux, the Vulkan SDK are split into dedicated packages
as recommended by LunarXchange [1].

Specifically, the two packages vulkan-headers and vulkan-icd-loader
contain the Vulkan headers and share libraries, respectively.

	$ pacman -Ql vulkan-headers
	vulkan-headers /usr/include/vulkan/vulkan.h
	vulkan-headers /usr/include/vulkan/vulkan.hpp

	$ pacman -Ql vulkan-icd-loader
	vulkan-icd-loader /usr/lib/libvulkan.so

As noted above, neither the Vulkan headers nor the shared library are
stored in a file path with a Vulkan SDK prefix directory. As such, the
CMake file prior to this commit did not support building VGE on
Arch Linux.

With this commit, building on Windows using CMake would require
setting VULKAN_SDK. As the previous setting required the SDK to be
placed on D:, which may not be the case for all Windows machines
this should make compilation on Windows more explicit.

From [1]: https://vulkan.lunarg.com/sdk/home#linux

> For other distributions of Linux, we recommend that you search your distribution's package repository for vulkan development libraries. Otherwise you may download and build from publicly available repositories.
>
> Loader: https://github.com/KhronosGroup/Vulkan-Loader
> Validation Layers: https://github.com/KhronosGroup/Vulkan-ValidationLayers
> GFXReconstruct: https://github.com/LunarG/gfxreconstruct
> Tools source: https://github.com/LunarG/VulkanTools
> Samples source: https://github.com/LunarG/VulkanSamples